Visuals
- New external models for Mi-24, Mi-17, Mi-6, CH-53, AH-1W, Ka-29;
- Model of CH-47 is improved and has new texture;
- Main rotor blades animation;
- Accurate rotor blades damage animation;
- Static main rotor blades doesn’t disappear when rotor RPM value is high and always visible with blurred rotors or without. To avoid “freezing” rotors rotating speed limited to variable value, linked to current framerate;
- New undercarriage animation mostly for all helicopters (except AH-64 and CH-3);
- Some sort of suspension for AI controlled helicopters - wheels are align to ground surface, but if maximum or minimum limit of current gear is reached it doesn't works;
- Blurred rotors became invisible if game paused;

User interface

- Details and status bar of current helicopter repairing and refueling while landed on allied keysite;

Collisions

- Suspension works while landed on carrier;
- Additional type of collision points added – Stabilizer, Engine and Avionics. Cause corresponding damage;
- Additional points are added to helicopter objects;
- Position of collision points are calculating every frame, so collision of moving parts are more accurate;
- Collisions calculating are changed a bit – it’s more precise and do not cause false clashes, but on other hand it works worse with very small objects;
- Collision of moving points (rotor blades) do not stop model but create dynamic force in opposite direction;
- Collision points of main rotors are moving automatically from damaged blade section to previous;
- Damaged undercarriage will still work but has less effectiveness;
- Suspension dynamics added for all helicopters;
- Additional undercarriage point types are added for CH-47 and Ka-29;
- Suspension configs moved to \common\data\suspension folder;
- Interaction of helicopter with water are added;
- Missing collision objects are added for Mi-28, Ka-52, AH-64D, AH-64A, AH1-Z, AH1-W, AH1-T, OH-58;

Damage

- Tail rotor damage enabled;
- Severity of Main rotor/Tail rotor/Stabilizer/Main rotor blade damage is decreased, Engine damage – increased;
- Delay of damage repairing while landed on allied keysite is fixed;
- Value of damage repairing delay is different every time;
- Damage can be taken both from land and other object collision;
- Water random damage is added – than longer helicopter staying on water surface than more damage it gets (except CH-47 and CH-53). Engines will be damaged immediately in case of any engine collision with water;
- Repairing cheat has usual delay when difficult is HARD;

Flight model

- Arneh’s flight model not switchable anymore, it works always. You can turn on Werewolf’s aerodynamics modifiers by setting “flight_model = 1”;
- Reaction force decreased while tail rotor is damaged;
- New damaged rotor vibration;
- Damaged stabilizer doesn’t affects dynamics while hovering or low speed flying;
- Weapon launch force is enabled. Little glitchy – just for testing and future adjustment;
- Main rotor spin-up vibrations;
- Sideway resistance of landed helicopter is increased;

Changes in GWUT file: Center of gravity and Fuselage angle updated for all helicopters, Main rotor radius fixed for MV-22 and CH-46.

+++ aphavoc/source/entity/system/en_attrs/en_attrs.h
@@ -117,7 +117,7 @@ extern entity_attribute_data
 
 #define ENTITY_ATTR_END entity_attr_end
 
-#define ENTITY_ATTR_ATTITUDE_ANGLES(HEADING,PITCH,ROLL) entity_attr_attitude_angles, ((float) (HEADING)), ((float) (PITCH)), ((float) (ROLL))
+#define ENTITY_ATTR_ATTITUDE_ANGLES(HEADING,PITCH,ROLL) entity_attr_attitude_angles, ((double) (HEADING)), ((double) (PITCH)), ((double) (ROLL))
 
 #define ENTITY_ATTR_CHAR_VALUE(CHAR_TYPE,VALUE) entity_attr_char_value, ((char_types) (CHAR_TYPE)), ((char) (VALUE))
